refactor: simplify service tracing by removing hook plugin architecture
- Move apply_replica() and replica state handling to new replica_state.py module
- Remove TracingHook protocol and ServiceTracingHook plugin-based approach
- Simplify BoundReplicaMethod to delegate directly to service provider
- Remove global replica hook registry (register_replica_hook, unregister_replica_hook)
- Add get_service_bindings() and register_service_bindings() helpers for direct binding management
- Streamline ServiceCallExecutor removal and integrate directly with execution context
- Update graph tracing to reject replica calls without proper @node/@bind_service decoration
- Update examples documentation to reference replica_state.py

This refactoring removes unnecessary indirection from the tracing hook architecture while
maintaining the same service execution functionality through a more direct approach.
