+++ examples/stateful_replica_patterns.py
@@ -611,6 +611,7 @@ async def main():
     print("\n" + "=" * 70)
     print("For more information, see:")
     print("  - thinkagain/core/execution/replica.py")
+    print("  - thinkagain/core/execution/replica_state.py")
     print("  - thinkagain/core/traceable.py")
     print("  - examples/distributed_example.py")
     print("  - tests/test_distributed.py")
