+++ thinkagain/core/execution/replica.py
@@ -5,9 +5,7 @@ from __future__ import annotations
 import cloudpickle
 import uuid
 from dataclasses import dataclass, field
-from typing import Any, Iterable
-
-from .runtime import maybe_await
+from typing import Any
 
 
 # ---------------------------------------------------------------------------
@@ -124,7 +122,7 @@ class BoundReplicaMethod:
     """A bound __call__ method on a replica handle.
 
     This is returned when calling a ReplicaHandle (e.g., await llm(prompt)).
-    When called, it delegates to the registered tracing hook if one exists.
+    When called, it delegates to the mesh service provider during execution.
     """
 
     def __init__(self, handle: ReplicaHandle):
@@ -176,106 +174,6 @@ class BoundReplicaMethod:
             )
         provider = mesh.get_service_provider()
         return await provider.execute_service_call(self.handle, args, kwargs)
-
-
-# ---------------------------------------------------------------------------
-# Replica Hook Registry
-# ---------------------------------------------------------------------------
-
-
-_replica_hook: Any = None  # TracingHook protocol
-
-
-def register_replica_hook(hook: Any) -> None:
-    """Register a tracing hook for replica calls.
-
-    This allows the distributed module (or other runtimes) to intercept
-    replica method calls and record them appropriately during tracing.
-
-    Args:
-        hook: Object implementing the TracingHook protocol
-    """
-    global _replica_hook
-    _replica_hook = hook
-
-
-def unregister_replica_hook() -> None:
-    """Unregister the current replica hook."""
-    global _replica_hook
-    _replica_hook = None
-
-
-def _get_replica_hook() -> Any:
-    """Get the currently registered replica hook."""
-    return _replica_hook
-
-
-# ---------------------------------------------------------------------------
-# Replica State Helpers
-# ---------------------------------------------------------------------------
-
-
-def _iter_slots(cls: type) -> Iterable[str]:
-    for base in cls.__mro__:
-        slots = getattr(base, "__slots__", ())
-        if isinstance(slots, str):
-            slots = (slots,)
-        for slot in slots:
-            if slot not in ("__dict__", "__weakref__"):
-                yield slot
-
-
-def _copy_replica_state(target: Any, source: Any) -> None:
-    if type(target) is not type(source):
-        raise TypeError(
-            "Replica compose() must return the same class as the target replica."
-        )
-
-    if hasattr(target, "__dict__"):
-        target.__dict__.clear()
-        target.__dict__.update(source.__dict__)
-
-    for slot in _iter_slots(type(target)):
-        if hasattr(source, slot):
-            setattr(target, slot, getattr(source, slot))
-
-
-async def apply_replica(replica_obj: Any, fn: Any, *args, **kwargs) -> Any:
-    """Apply a @jit-compatible state update to a replica via decompose/compose.
-
-    The replica must implement:
-      - decompose(self) -> (children: list[Any] | tuple[Any, ...], aux: Any)
-      - compose(cls, aux, children) -> replica instance
-
-    The function must return (new_children, output), where new_children matches
-    the structure of decompose()'s children.
-    """
-    if not hasattr(replica_obj, "decompose") or not hasattr(
-        replica_obj.__class__, "compose"
-    ):
-        raise TypeError(
-            "Replica must define decompose() and compose() to use apply_replica()."
-        )
-
-    children, aux = replica_obj.decompose()
-    if not isinstance(children, (list, tuple)):
-        raise TypeError("decompose() must return a list/tuple of children.")
-
-    result = await maybe_await(fn, *children, *args, **kwargs)
-    if not (isinstance(result, tuple) and len(result) == 2):
-        raise TypeError("Expected (new_children, output) from replica update.")
-
-    new_children, output = result
-    if not isinstance(new_children, (list, tuple)):
-        new_children = [new_children]
-    if len(new_children) != len(children):
-        raise ValueError(
-            "Updated children count must match decompose() children count."
-        )
-
-    updated = type(replica_obj).compose(aux, list(new_children))
-    _copy_replica_state(replica_obj, updated)
-    return output
 
 
 # ---------------------------------------------------------------------------
