+++ thinkagain/core/protocols.py
@@ -25,26 +25,3 @@ class ServiceProvider(Protocol):
             Result of the service method call
         """
         ...
-
-
-class TracingHook(Protocol):
-    """Hook for recording operations during graph tracing.
-
-    This protocol allows external systems (like service runtime) to participate
-    in tracing without coupling the core tracing infrastructure to specific concepts.
-    """
-
-    async def record_call(self, args: tuple, kwargs: dict, handle: Any) -> Any:
-        """Record a traced operation for a service call."""
-        ...
-
-    def get_resource_index(self, resource: Any) -> int:
-        """Get or register a resource (e.g., service handle) in the trace.
-
-        Args:
-            resource: Resource to register
-
-        Returns:
-            Index where resource is stored in graph inputs
-        """
-        ...
